Fix AhoCorasickNode DebuggerDisplay
